fix(@xen-orchestra/restapi): updates from reviews
