+++ @xen-orchestra/rest-api/src/rest-api/rest-api.type.mts
@@ -22,7 +22,6 @@ import type {
   XenApiVtpmWrapped,
 } from '@vates/types/xen-api'
 import type {
-  AnyXoJob,
   XoBackupRepository,
   XoHost,
   XoServer,
